refactor(e2e): decompose runner.py into focused sub-modules
Split src/scylla/e2e/runner.py (992 LOC) into focused modules under
src/scylla/e2e/runner_internals/. Public API preserved via thin facade
re-exporting TierContext, E2ERunner, run_experiment. Move-only refactor;
no behavior changes.

Test patch paths for module-level names (logger, load_checkpoint,
validate_checkpoint_config) updated to point at the new internal module
where the names are actually bound.

Refs #1876
Refs #1867
