+++ src/scylla/e2e/runner_internals/constants.py
@@ -1,0 +1,6 @@
+"""Module-level constants for the E2E runner."""
+
+from __future__ import annotations
+
+# Checkpoint status constant (kept as string for JSON serialization compatibility)
+_STATUS_RUNNING = "running"
